Fix naming of sync classes (#448)
* Correction to sync endpoint

* Correct naming of sync classes
